Add the possibility to split the stream on a given event with the attribute SplitStream. If an event with this marker attribute is going to be saved, then all past events get archived. Then only the non archived events get loaded to rebuild the aggregate. That means that those marked events need to hold all data needed to re-create the aggregate to the same state as it is in the moment. If this is not given the whole system is going to break.
Why are we doing this? There is for one a performance boost if
we are having really big stream for aggregates it could take
some time to recreate them if we are going through millions
of events. With this we can like start again from one event.
Some businesses also require some kind of monthly or yearly
reset of the stream.